Add files via upload
The Purchasing Managers Index is a measure of the prevailing direction of economic trends in manufacturing and is used by Supply chain Managers in more than 19 industries. A reading of higher than 50 indicates an economic expansion. This means that the closer the reading is to 100, the higher the degree of positive economic growth. A reading below 50 indicates an economic contraction, with readings closer to 0 indicating a higher degree of contraction.